Narek edit 1: fix python bin determinner fix paths new logger using python lib
