+++ app/data/runner_log.txt
@@ -1,0 +1,40 @@
+2025-01-08 20:45:43,636 - INFO - --------STARTING RUNNER-------->
+
+2025-01-08 20:45:43,636 - INFO - Running SHELL_CMD=/usr/bin/python3 ./scripts/ChurchArmenia.py
+
+2025-01-08 20:45:43,658 - CRITICAL - 	ended with exit code <2>
+
+2025-01-08 20:45:45,659 - INFO - Running SHELL_CMD=/usr/bin/python3 ./scripts/ChurchArmenia.py
+
+2025-01-08 20:45:45,684 - CRITICAL - 	ended with exit code <2>
+
+2025-01-08 20:45:47,684 - INFO - Running SHELL_CMD=/usr/bin/python3 ./scripts/ChurchArmenia.py
+
+2025-01-08 20:45:47,710 - CRITICAL - 	ended with exit code <2>
+
+2025-01-08 20:45:49,711 - INFO - Running SHELL_CMD=/usr/bin/python3 ./scripts/ChurchArmenia.py
+
+2025-01-08 20:45:49,737 - CRITICAL - 	ended with exit code <2>
+
+2025-01-08 20:45:51,737 - INFO - Running SHELL_CMD=/usr/bin/python3 ./scripts/ChurchArmenia.py
+
+2025-01-08 20:45:51,765 - CRITICAL - 	ended with exit code <2>
+
+2025-01-08 20:45:52,310 - INFO - --------ENDING RUNNER 	 KEYBOARD INTERRUPT--------
+
+2025-01-08 20:45:52,310 - INFO - --------EXITING RUNNER SCRIPT - BYE--------
+
+2025-01-08 20:46:56,836 - INFO - --------STARTING RUNNER-------->
+
+2025-01-08 20:46:56,837 - INFO - Running SHELL_CMD=/usr/bin/python3 /home/debian/Desktop/Mnvsss/ChurchArmenia_tg_bot/app/scripts/ChurchArmenia.py
+
+2025-01-08 20:46:56,860 - CRITICAL - 	ended with exit code <1>
+
+2025-01-08 20:46:58,860 - INFO - Running SHELL_CMD=/usr/bin/python3 /home/debian/Desktop/Mnvsss/ChurchArmenia_tg_bot/app/scripts/ChurchArmenia.py
+
+2025-01-08 20:46:58,884 - CRITICAL - 	ended with exit code <1>
+
+2025-01-08 20:47:00,212 - INFO - --------ENDING RUNNER 	 KEYBOARD INTERRUPT--------
+
+2025-01-08 20:47:00,213 - INFO - --------EXITING RUNNER SCRIPT - BYE--------
+
